pyproject.toml
[project]
name = "aprs2mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "latloncalc>=1.5.6",
    "paho-mqtt>=2.1.0",
    "pytest>=8.3.5",
]
